segmentation/requirements.txt
absl-py==1.2.0
albumentations==1.3.0
asgiref==3.5.2
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
bleach==6.0.0
bunch==1.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==37.0.4
cssselect==1.1.0
cv==1.0.0
cycler==0.11.0
decorator==5.1.1
Django==4.2
django-janitor==0.5.1
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
executing==1.2.0
filelock==3.7.1
Flask==2.2.3
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
guided-filter-pytorch==3.7.5
h5py==3.7.0
hyperlink==21.0.0
idna==3.3
imageio==2.25.1
imgaug==0.4.0
incremental==21.3.0
ipython==8.12.0
itemadapter==0.6.0
itemloaders==1.0.4
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonpickle==3.0.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
loguru==0.7.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
ml-collections==0.1.1
munch==2.5.0
network==0.1
networkx==3.0
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parsel==1.6.0
parso==0.8.3
pathtools==0.1.2
pickleshare==0.7.5
Pillow==9.2.0
pretrainedmodels==0.7.4
prettytable==0.7.2
promise==2.3
prompt-toolkit==3.0.38
Protego==0.2.1
protobuf==3.19.4
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydicom==2.3.1
PyDispatcher==2.0.5
pyecharts==2.0.2
Pygments==2.14.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pyvis==0.3.2
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Scrapy==2.6.2
segmentation-models-pytorch==0.3.1
sentry-sdk==1.12.1
service-identity==21.1.0
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
simplejson==3.17.0
six==1.16.0
smmap==5.0.0
soupsieve==2.0
sqlparse==0.4.2
stack-data==0.6.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.2.3
timm==0.4.12
tldextract==3.3.1
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
traitlets==5.9.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.25.8
w3lib==1.22.0
wandb==0.13.7
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.2.3
win32-setctime==1.1.0
wrapt==1.14.1
zope.interface==5.4.0
